# Replace System.Random with Xorshift RNG.