--- a/Benchmark/WrappedInt.cs
+++ b/Benchmark/WrappedInt.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Benchmark
+{
+    struct WrappedInt : IEquatable<WrappedInt>
+    {
+        public int Value;
+
+        public bool Equals(WrappedInt other) => this.Value == other.Value;
+
+        public override int GetHashCode() => 42;
+
+        public override bool Equals(object obj) => this.Equals((WrappedInt)obj);
+
+        public static implicit operator WrappedInt(int i) => new WrappedInt { Value = i };
+    }
+}